requirements.txt
keras
numpy
pandas
python-dotenv
requests
streamlit
tensorflow